app, libgimp, pdb: use 64-bit window id
This is a follow-up of the clang issues signaled in #8649. When testing
CLANG64 on Windows under MSYS2, there was another similar issue in
gimpwidgets-utils.c.
In addition to that, the previous cast using GPOINTER_TO_INT proved
incorrect, since that discards anything except the lower 32-bits.
Since we already return guint64 here, we will cast this to guint64 too.

By changing the return to guint64 in gimpwidgets-utils.c too, we kind of
snowball into needing to change multiple functions that so far used
guint32 for window ids.

+++ app/core/gimppdbprogress.c
@@ -74,7 +74,7 @@ static void     gimp_pdb_progress_progress_set_value     (GimpProgress *progress
                                                           gdouble       percentage);
 static gdouble  gimp_pdb_progress_progress_get_value     (GimpProgress *progress);
 static void     gimp_pdb_progress_progress_pulse         (GimpProgress *progress);
-static guint32  gimp_pdb_progress_progress_get_window_id (GimpProgress *progress);
+static guint64  gimp_pdb_progress_progress_get_window_id (GimpProgress *progress);
 
 
 static GObjectClass *parent_class = NULL;
@@ -376,12 +376,12 @@ gimp_pdb_progress_progress_pulse (GimpProgress *progress)
                                     NULL, 0.0);
 }
 
-static guint32
+static guint64
 gimp_pdb_progress_progress_get_window_id (GimpProgress *progress)
 {
   GimpPdbProgress *pdb_progress = GIMP_PDB_PROGRESS (progress);
 
-  return (guint32)
+  return (guint64)
     gimp_pdb_progress_run_callback (pdb_progress,
                                     GIMP_PROGRESS_COMMAND_GET_WINDOW,
                                     NULL, 0.0);
