+++ app/widgets/gimpwidgets-utils.h
@@ -71,9 +71,9 @@ gboolean          gimp_get_style_color             (GtkWidget            *widget
                                                     GdkRGBA              *color);
 void              gimp_window_set_hint             (GtkWindow            *window,
                                                     GimpWindowHint        hint);
-guint32           gimp_window_get_native_id        (GtkWindow            *window);
+guint64           gimp_window_get_native_id        (GtkWindow            *window);
 void              gimp_window_set_transient_for    (GtkWindow            *window,
-                                                    guint32               parent_ID);
+                                                    guint64               parent_ID);
 void              gimp_widget_set_accel_help       (GtkWidget            *widget,
                                                     GimpAction           *action);
